added support for abstract class

+++ include/miyuki/hw/texture.h
@@ -9,23 +9,23 @@
 
 namespace Miyuki {
 	namespace HW {
-		class Texture : NonCopyable{
+		class Texture : NonCopyable {
 			GLuint texture;
 			size_t width, height;
 		public:
 			GLuint getTexture()const { return texture; }
 			Texture(size_t width, size_t height);
-			Texture(size_t width, size_t height, uint8_t * data);
+			Texture(size_t width, size_t height, uint8_t* data);
 			Texture(const IO::Image&);
-			Point2i size() const{ return Point2i(width,height); }
+			Point2i size() const { return Point2i((int)width, (int)height); }
 
 			// clears all contents
 			void resize(size_t width, size_t height);
 
-			void setPixel(int x, int y, const Spectrum& color);
+			void setPixel(size_t x, size_t y, const Spectrum& color);
 			void setPixel(const Point2f& p, const Spectrum& color) {
-				int x = clamp<int>(p[0], 0, width - 1);
-				int y = clamp<int>(p[1], 0, height - 1);
+				size_t x = clamp<size_t>(p[0], 0, width - 1);
+				size_t y = clamp<size_t>(p[1], 0, height - 1);
 				setPixel(x, y, color);
 			}
 			void setData(uint8_t* data);
@@ -33,6 +33,6 @@ namespace Miyuki {
 			void use();
 
 			~Texture();
-		};		
+		};
 	}
 }
